requirements.txt
# DeepSearchAgent dependencies
uv>=0.4.0
smolagents>=1.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
litellm>=1.40.0
pydantic>=2.5.0
toml>=0.10.2
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
aiofiles>=24.1.0

# Additional dependencies for apartment search
beautifulsoup4>=4.12.0
lxml>=5.1.0
selenium>=4.15.0